% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_legend_helper_functions.R
\name{annotate_legend}
\alias{annotate_legend}
\title{Annotate a ggplot legend with two labels on either side}
\usage{
annotate_legend(plot, label1, label2)
}
\arguments{
\item{plot}{A ggplot object}

\item{label1}{The text for the first (left) label}

\item{label2}{The text for the second (right) label}
}
\value{
A ggplot object with the modified legend
}
\description{
Adapted from: https://stackoverflow.com/a/37235413
}
