% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_legend_helper_functions.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Create a shared leged (with annotations) for multiple plots}
\usage{
grid_arrange_shared_legend(..., ncol = 1, nrow = length(list(...)),
  position = c("bottom", "right"))
}
\arguments{
\item{...}{two or more ggplot objects}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{position}{position of the legend}
}
\value{
A ggplot object with the combined plots
}
\description{
Adapted from: https://stackoverflow.com/a/38420690 and https://stackoverflow.com/a/37235413
}
